fix installing dnsmasq on Ubuntu 16.
Aparently Ubuntu 16 ships with a copy of dnsmasq but it's not installed, and checking if the dnsmasq command being available or not is not a reliable way to determine if dnsmasq is installed or not. So the exit code of the dnsmasq service status is used. If anything other than 0 is returned, the installer tries to install dnsmasq.