Use service or systemctl to check if dnsmasq is working
Not all systems use systemctl, not all will have service in the future. This setup tries to use service if it's available to check the status of the dnsmasq service (to determine whether to try to install it or not). If service is unavailable, it then checks if systemctl is available or not.