fogsettings gets updated, 1 NIC is ok now
Cleaned out some unused functions and wrote code to update the /opt/fog/.fogsettings file and it's relevant fields that may be concerned with a NIC/IP change. Also wrote code to correct for only having one interface because because previously the script would exit if there was not more than one.