one minute between deploy starts
This is because deployments were getting too fast. One of the recent deployments actually finished before the other was even queued up. This is mostly due to the SSD disk upgrade that these guests run on - and also due to switching to ZST compression - and just fog being so fast to begin with.